fix #34
In case of marking lines series data and config.l are not in synch because series data uses a marking-offset for marking line data which is stored at the beginning of the series data. A corresponding config.l does not exist at this position. Using the global marking offset for accessing the correct series data in case of bar labels solves the issue.